+++ include/interface/platform/delay/stm32/delay_if_stm32h7xx.hpp
@@ -22,16 +22,16 @@ using Impl = platform::delay::DelayImpl<Backend>;
 
 EMPP_STATIC_INLINE void init() EMPP_NOEXCEPT { Impl::init(EMPP_SYSCLK_MHZ); }
 
-EMPP_STATIC_INLINE void init(const uint16_t sysclk_mhz) EMPP_NOEXCEPT
+EMPP_STATIC_INLINE void init(const unsigned int sysclk_mhz) EMPP_NOEXCEPT
 {
     Impl::init(sysclk_mhz);
 }
 
-EMPP_STATIC_INLINE void us(const uint32_t nUs) EMPP_NOEXCEPT { Impl::us(nUs); }
+EMPP_STATIC_INLINE void us(const size_t nUs) EMPP_NOEXCEPT { Impl::us(nUs); }
 
-EMPP_STATIC_INLINE void ms(const uint16_t nMs) EMPP_NOEXCEPT { Impl::ms(nMs); }
+EMPP_STATIC_INLINE void ms(const size_t nMs) EMPP_NOEXCEPT { Impl::ms(nMs); }
 
-EMPP_STATIC_INLINE void s(const uint16_t nS) EMPP_NOEXCEPT { Impl::s(nS); }
+EMPP_STATIC_INLINE void s(const size_t nS) EMPP_NOEXCEPT { Impl::s(nS); }
 
     #if EMPP_DELAY_USE_DWT
 /* 测量运行时间 */
